+++ client/src/App.js
@@ -8,13 +8,7 @@ import Navbar from "./components/navbar";
 import RecordList from "./components/recordList";
 import Edit from "./components/edit";
 import Create from "./components/create";
-import DiningCard from "./components/dinningCard";
-import Col from 'react-bootstrap/Col';
-import Row from 'react-bootstrap/Row';
-
-import cardData from "./cardData";
-// import "./components/cardindex.css"
-
+import DiningCard from "./components/diningCard";
 
 import Home from './pages/home';
 import Dining from './pages/dining';
@@ -38,14 +32,6 @@ const App = () => {
             </Routes>
             <div className="home-cards">
                 <DiningCard />
-                {/* <DiningCard deatails={cardData} /> */}
-                {/* <Row xs={1} md={2} className="g-4">
-                {Array.from({ length: 3 }).map((_, idx) => (
-                    <Col>
-                        <DiningCard deatails={cardData} />
-                    </Col>
-                ))}
-                </Row> */}
             </div>
             
         </div>
