+++ client/src/components/diningCard.js
@@ -1,0 +1,44 @@
+import React, { useEffect, useState} from 'react';
+import { Card, Row, Col, Container } from "react-bootstrap";
+import "./diningCard.css"
+
+function DiningCard() {
+    const [diningData, setdiningData] = useState([]);
+    useEffect(() => {
+        const fetchData = async () => {
+            const response = await fetch(`http://localhost:8000/dining/`);
+            
+            if (!response.ok) {
+                const message = `An error occurred: ${response.statusText}`;
+                window.alert(message);
+                return;
+            }
+
+            const records = await response.json();
+            setdiningData(records.result);
+        }
+        fetchData();
+    },[])
+  return (
+    <Container>
+        <Row guter={40} className="row">
+        {diningData.map((value, index) => (
+        <Col key={index} xs={12} md={6} lg={6}>
+             <Card >
+                 <Card.Img src={value.image} alt="dinning images" />
+
+                 <Card.Body>
+                     <Card.Title>{value.name} <span class="badge bg-secondary">{value.onCampus?"On Campus":"Off Campus"}</span></Card.Title>
+                     <Card.Text>{value.location}</Card.Text>
+                 </Card.Body>
+             </Card>
+         </Col>
+            
+        ))}
+        </Row>
+    </Container>
+    
+  );
+};
+
+export default DiningCard;
